Fix APU and CPU parity (#15)
* Fix memory allocations

* Fix Controller allocation

* Fix PPU and APU

* Add allocations test

* Bump compose and gradle

* Remove serializable tags

* Cleanup

+++ frontend/src/commonMain/kotlin/frontend/SharedFrameBuffer.kt
@@ -4,14 +4,14 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 
 class SharedFrameBuffer {
-    private val blankFrame = ByteArray(256 * 240)
+    private val blankFrame = IntArray(256 * 240)
     private val _frameTicks = MutableStateFlow(0L)
 
-    var currentFrame: ByteArray = blankFrame
+    var currentFrame: IntArray = blankFrame
         private set
     val frameTicks = _frameTicks.asStateFlow()
 
-    fun submit(framebuffer: ByteArray, frameCount: Int) {
+    fun submit(framebuffer: IntArray, frameCount: Int) {
         currentFrame = framebuffer
         _frameTicks.value = frameCount.toLong()
     }
